Add infrastructure for generating OpenGL bindings
Our Tetris clone will be OpenGL based (though it is conceivable that
other backends will become available in the future). Contrary to C,
there is no system-provided OpenGL header to include and instead all
crates have opted to generate bindings at one point or another. We plan
on using a fairly dated version of OpenGL, which will always be
available. As such, dynamic loading of those bindings is unlikely to add
anything.
With this change we generate a static set of OpenGL bindings and
introduce the infrastructure for recreating them should the need ever
arise (e.g., if gl_generator experiences major changes) in the form of a
feature switch.