Introduce basic windowing infrastructure
Being OpenGL based (as opposed to, say, terminal based), we need a
proper window with an OpenGL context into which we can render. With this
change we introduce the necessary infrastructure for going about the
creation of it. We rely on the de-facto standard crate for this purpose
in Rust, which is winit.