Introduce OpenGL renderer
All our low-level rendering details (and OpenGL specifics in particular)
are meant to be abstracted away from the core game logic, so that we can
more easily swap out the rendering backend to use.
The way we intend to go about that is by using a dedicated "renderer" to
funnel all render operations through.
This change introduces and uses a first version of this type, which
currently just makes sure to clear the window background with a uniform
color.