Set up OpenGL state and initialize view
With this change we make sure to configure OpenGL as the game will need
it, by disabling functionality we do not need, enabling things we will
use, and setting up matrizes and the viewport etc. We also handle resize
events to make sure that our view is updated in accordance with the new
window size.