Introduce color guard for renderer state management
Our Renderer type is stateful, in part to mirror how OpenGL works as
well. That is to say, you set a certain color on it and all subsequent
rendering operations will make use of it, until changed further.
However, once rendering becomes more complex and/or conditional, just
setting a color and forgetting about it can quickly become a mess:
subsequent code may see different states based on what functionality
executed earlier. A similar argument will apply to textures and other
state the renderer will deal with in the future.
To make state management more predictable, this change adjusts the
Renderer type's state changing methods to return a Guard object, which
contains the previous state and will make sure to restore said state
once it goes out of scope. In so doing we introduce more predictability
into the rendering pipeline, which will ultimately allow for rendering
code to make more assumption about what state is already in effect (and,
thus, alleviate the need for re-setting it).