Add support for creating and rendering with textures
This change introduces the means for creating textures and using them
when rendering primitives. We use the image crate for loading and
decoding the image data and then map different common formats to
something that OpenGL understands. Because so far we only have the
ability to render lines and textured lines don't make too much sense,
this change also add support for rendering rectangles.