Introduce concept of ActiveRenderer
Currently there is no guarantee that a Renderer::on_pre_render() call is
made before any actual rendering takes place. Similarly, nothing really
says that Renderer::on_post_render() should be called afterwards.
These are invariants that lend themselves to being expressed in Rust's
type system, which makes misuse a compile time error.
With this change we do exactly that --- we introduce an ActiveRenderer
which can be used to render primitives. An object of this type can only
be obtained by calling the on_pre_render() method on a Renderer object
itself. Similarly, once the ActiveRenderer goes out of scope it will
take care of calling on_post_render().
Note that as one side-effect of invoking on_post_render() as part of a
destructor, we loose the ability to bubble up errors on this path, such
as from the swap_buffers() call. We opted for aborting the program
should such an error occur, as that seems more apt than merely logging
somewhere. In general we expect a failure of this call to be due to a
bug anyway.