Bind textures more lazily
The setting of a texture (along with other state) to use during
rendering always results in a guard object being created that restores
the previously active texture. Currently, this restoration eagerly binds
the previous texture in question. However, there is a reasonably high
likelihood that this texture won't actually be used for any rendering.
That would mean that we bound it completely unnecessarily.
To prevent such needless state changes from the get-go, this change
makes the restoration of a texture only remember it, but not actively
bind it with OpenGL. Only once it is actually used will it be fully
bound.