Introduce concept of a rendering origin to renderer
We will render different components of the game at different locations
on the window. But while our view of the window is basically of a fixed
size of game units, we may still want to treat some game components as
being located relative to another. In that sense, some components act as
the origin relative to which rendering operations are to happen.
This could obviously simply be expressed at the component level, by
adding the two coordinates. However, given our stateful rendering
approach, it is perhaps easiest to just have the renderer carry the
notion of an origin and have it adjust all primitives relative to this
origin.
With this change we add the necessary logic for doing just that.