Add support for clearing the currently active texture
In general our assumption is that most code will set a texture and do
its rendering. However, sometimes the intention is to have no texture in
use, irrespective of what previous code did. We don't currently support
that use case with public APIs.
With this change we introduce the ActiveRenderer::set_no_texture() call,
which does precisely that. Internally we just make sure to keep around
an invalid texture and with that we set just set it as if it was an
actual texture to achieve the desired outcome.