Introduce main Game object with Tetris field
This change introduces the main Game type that will hold all Tetris
specific logic and instantiates it. Right now the game only contains the
empty field, which includes a background image as well as the side
walls. These components are textured and the textures are embedded into
the binary at compile time, to avoid the need for shipping "assets"
along the game.