Move Renderer out of Window type
So far our Renderer type was part of a Window. While that makes sense at
a conceptual level, it poses an interesting challenge: the renderer is
the entity that, among many other things, maintains the view, expressed
in logical (or game) coordinates. In order to calculate this view, it
needs both the physical window size as well as the desired logical size.
Currently that is not a problem, because the logical size is just hard
coded. However, in the future we want to calculate the dimensions based
on the bounds of the components in the game. To do that, in turn, it
is necessary to instantiate the game. However, that results in a
circular dependency: we can't create the game without having a window
(because it requires the loading of textures which will fail without
OpenGL being initialized properly). But we need a game object to
instantiate a Renderer (to know desired view dimensions).
To break this cycle, this change moves the Renderer out of the Window
type. In so doing we will be able to instantiate the window without
caring for game dimensions. That will allow us to create a Game object,
which subsequently provides the prerequisite data to instantiate a
Renderer.