Use actual game dimensions when instantiating Renderer
So far we have instantiated the Renderer with hard coded dimensions of
the game. That was fine and necessary, because we didn't have any live
dimensions available.
By now we have a Game object that knows the width and height that the
game requires and with this change we provide these dimensions to the
Renderer.