Introduce pseudo random number generator
In the future we want to generate random Tetris stones. Rust's standard
library does not contain functionality for creating random numbers that
we could use for this purpose.
Hence, with this change we create our own pseudo random number generator
based on a permuted congruential generator. By default, we seed it with
the system time, which will ensure that different instantiations yield
different random sequences.