Introduce Stone type
This change introduces the next necessary component of the Tetris game:
the stone. A stone consists of any number of "pieces" (traditionally
four; a tetromino). Pieces are very lightweight, carrying only their
color. Location and texture are effectively externally managed. Stones
can ultimately be destructed into the individual pieces, which will be
what is stored in the field once a collision with already dropped pieces
is detected, but that logic does not yet exist. Stones can be moved,
which really just moves all the individual pieces.