Introduce stone factory
As motivated earlier, in order to play the game we are going to need
Tetris stones. Now that we have the concept of a Stone as well as the
means to generate random numbers, this change marries those two into a
stone factory, which can be used to create random stones based on stone
templates. We include templates for the traditional tetrominos (stones
containing four pieces).