Create and position stones in the Tetris field
With the recently added stone factory we have the means for creating
random Tetris stones.
With this change we incorporate that functionality into the Field type,
which will ensure to have one (user controllable) stone available at the
expected position while the game is active. We abstract stone creation
behind the newly introduced StoneProducer trait to abstract from the
underlying creation mechanism. We anticipate switching out the factory
with another producing component that keeps a queue of stones that will
also be displayed in a preview area.