Correctly trigger re-render on state changes
So far we blindly relied on the RedrawRequested system event for
triggering a re-render of the game. However, in and of itself this event
is only emitted for changes that the windowing system can be aware of.
We want to render the scene whenever something changed inside the game,
however.
To make that happen we introduce the State enumeration and make all
paths of the main event loop emit whether something changed and requires
a redraw (State::Changed) or not (State::Unchanged). If anything reports
a change, we queue a RedrawRequested event by informing the window about
the desire to redraw.
Note that it is a deliberate choice of ours that we render on-demand
only and not continuously as many more complex games do. Rendering only
when something changed makes much more sense for a game such as Tetris,
that does not exhibit constant movement but is entirely driven by
"ticks". In such cases, on-demand rendering greatly helps in reducing
CPU and GPU usage (and with that power consumption caused by the game).