Add collision detection support for stones
We need a way to detect when a stone collides with a wall or with other
already dropped pieces, in order to either prevent the stone's movement
(so that it won't pass through a wall) or to "merge" it with the field
and trigger the creation of a new stone.
This change introduces the means for detection such a collision. Given
the way we represent a stone, all that involves is checking for overlap
between the locations of the stone's individual pieces with the wall or
with other already merged pieces.