Add support for merging a stone with the field
Now that we can detect a collision between a stone and already dropped
pieces in the field, we can take it one step further and merge the stone
with the field if such a collision happened as part of a downward move.
In such an event, we use the stone producer to create a new stone.