Disable vsync
The recently added benchmark has basically confirmed what we already
suspected: that vsync is enabled by default and that is effectively
capping "rendering" performance, because buffer swaps will be
synchronized with video frames.
We don't want such a behavior. We will be rendering on-demand only (as
opposed to constantly at the maximum possible frame rate) and so this is
an artificial slow down that impacts user experience because it delays
handling of events while rendering is blocked. To that end, this change
disables vsync.