Add support for dropping a stone
This change introduces the means for "dropping" a stone to the game (or
more precisely: the field). Dropping a stone refers to moving it down so
far that it collides with the field and is merged into it, resulting in
a new stone being created. The functionality is currently controlled by
the space key.