Add tick-based downward Stone movement
This change introduces automatic tick-based downward movement of the
currently active stone to the game. Movement speed will ultimately be
based on the game's current level, but because we still lack the notion
of a level it is currently fixed.
Movement is tick-based, meaning that we calculate the time at which the
next movement happens and only wake up again then. That is in opposition
to commonly used techniques where we accumulate minuscule stone
movements at the rendering frame rate and once it reaches and integer
value we adjust what is being displayed. Tick-based movement is deemed
superior for our intents and purposes, because it has the potential to
greatly reduce CPU usage.