Support auto-repeat for multiple keys
Owing to the way winit processes key events, we will never receive
auto-repeated key presses for multiple keys. That is due to the fact
that once a key is pressed, any currently active auto-repeats are
effectively canceled.
Such behavior is obviously a problem in many gaming contexts, where
movement can happen based on multiple keys pressed at the same time.
Because even something as basic as Tetris effectively requires this
behavior, this change introduces an abstraction that provides us with
the desired key-repeat functionality.
In a nutshell, we remember pressed keys until we see the corresponding
release event. We then query the system's default auto-repeat timeout
and interval and apply these for emitting "virtual" key press events for
each key still marked as pressed.