Properly detect game-over conditions
Up to this point, once the user had so many pieces in the field that no
new stone could be created without an immediate collision, we hit an
assertion failure and aborted the program.
That behavior was obviously never meant to stay that way and with this
change we make sure to gracefully end the game upon encountering this
condition instead. In many ways "game-over" is similar to "paused", with
the exception that it cannot be recovered from currently.