Add support for restart the game
Now that the game can end gracefully once it is no longer possible to
place a newly created stone without immediate collision, we need a way
to restart the game.
This change adds the necessary logic for doing so. Restarting the game
is currently bound to the Enter/Return key.