Keep track of score
Now that we can actually detect completed lines and clear them, it makes
sense for us to also keep score of how many lines were cleared.
This change introduces the logic for keeping track of cleared lines. On
top of that, we now have the means to properly increase the game's level
(when sufficient lines have been cleared) and to keep track of points
(with more points being awarded for more lines being cleared with a
single stone).
We currently don't have a score board (for lack of means to render
text), so right now we just print the level and points to standard
output when the level changes.