Work with physical rather than logical keys
When we switch from a different window to the Tetris window using
Meta+L, we see a key press event for both keys. However, only Meta
("Super") is ever reported as being released again. The reason for this
behavior is actually not entirely clear. It turns out, though, that we
actually do get a release event, but only for the physical L key and in
the form of a device event, not the logical one in a window event. While
this may or may not be a bug in winit itself, we don't care enough about
the difference to live with this broken behavior.
As such, this change switches us over to working with physical keys in
device events instead of logical ones in window event, to fix the
aforementioned problem.