Use i16 to represent game coordinates (and bounds)
So far we have used u16 for representing all game coordinates. That has
been mostly fine, except in some cases where we potentially (and
temporarily) moved stones or pieces left or below of the origin. In the
past we have relied on wrapping arithmetic to correctly handle those
cases, however, that can lead to tricky problems when callees do not
handle such wraps correctly or in the unlikely event that a wrap results
in a valid coordinate after all.
Given the nature of these issues, relying on signed arithmetic instead
seems like the saner choice and so with this change we switch all game
coordinates from u16 to i16. In addition, because coordinates and bounds
(heights & widths) are often intermixed, we represent those in the same
manner.