Add support for previewing upcoming stones
It can help players tremendously to be able to peek at upcoming stones
in order to make placement decisions.
With this change we add support for showing a preview of a configurable
number of upcoming stones. We do so in the form of a new PreviewStones
type. This type implements the StoneProducer trait and so it can be
transparently swapped in to be used by the Field for stone creation
matters.