Allow configurability via config file
The various default settings, while sensible, are not everybody's cup of
tea. Details such as the start level or the number of lines to clear in
order to level up have a tremendous impact on the game play experience.
With this change we make certain aspect of the game configurable using
the $XDG_CONFIG_DIR/tetris/config.toml configuration file. We print the
default configuration as part of the help text, to make it obvious what
a user would put in said file.