Make keyboard settings configurable
With the introduction of the configuration file we made various game
related state configurable. What is not covered, however, is the
keyboard auto-repeat configuration. While the system defaults are
entirely sensible as a starting point, users may reasonably expect to be
able to fine tune these settings.
With this change we expose knobs for the keyboard auto-repeat timeout
and interval via the configuration file.