Introduce font for rendering strings
So far we have not been able to render text to the screen. There are
libraries out there that can help us with that, but in the interest of
keeping a small foot print while providing a fairly custom
look-and-feel, with this change we introduce our own font that can be
used for rendering strings.
The font is based on a raster/bitmap font that exists in the form of a
bunch of bytes in source code, so it can be trivially adjusted if need
be. That means it won't offer TrueType-style looks, but that is by
design: each "pixel" in the bitmap actually maps to one textured
rectangle in our font. In so doing we can easily make any text look like
it is comprised of Tetris pieces itself (though potentially scaled).