Render score board to screen
So far we have only printed the score to stdout (and only on level
change, to not spam the terminal). That's obviously not the greatest
experience.
With this change we use the recently introduced font to render a proper
"score board" to the screen. The data will be rendered right of the game
but below the preview stones. Note that the size of the board is not
taken into account when calculating the overall "game view". As a
result, the score board may not be fully visible depending on window
dimensions. The reason for this behavior is that we would like to
specify the font size in pixels (or certainly something having a
relationship to the window size; as opposed to just using game
coordinates). Basically by design, all dimensions currently used in the
game can be expressed statically (either in absolute or relative
manner), in game coordinates. Because we rely on window size, which is
only available when rendering, we cannot size the overall game view
statically based on the font size.