Rework Game::toggle_pause() functionality
With an upcoming change we would like to pause and unpause the game when
the containing window looses/gains focus. The current
Game::toggle_pause() API is unfortunately not suitable to implement such
logic.
To that end, this change splits the method into two: pause() can be used
to pause/unpause the game using a flag argument and is_paused() returns
whether the game is currently paused.