Reduce the number of texture binds
In the past we already made binding of textures lazy, but that extended
only to the "restore-path". The reasoning there was that if we restore a
previously active texture, it is unlikely that said texture will be used
for actual rendering operations and so we just postpone actually binding
it in the graphics device until rendering operations are queued.
With this change we take things one step further still, by remembering
which texture is actually bound in the graphics device. If we end up
"restoring" a texture, then never use it, and then "restore" the same
texture as is already bound, we don't really have to do anything. This
is actually a very common occurrence: we start off with the invalid
texture, draw a stone, which activates the piece texture, then we
restore the invalid texture once we are done rendering the stone, then
we move on to the field, which uses the piece texture again, which is
now detected as still being bound. In so doing we also streamline the
code a bit, because binding a texture is now reduced to a single
location.
It is strictly speaking fairly easy for the graphics sub-system to
already optimize this case, but it's unclear how different OpenGL
implementations/drivers behave and so it makes sense to have this logic
reside on our end.