Allow for disabling key repeat
Not all keys we handle are directly involved in game play. For those
that are not, having a potentially very high repeat rate (along with a
short repeat timeout) may be detrimental to user experience. For
example, when we try to pause the game using F3, it frequently happens
that we press the key for a bit too long and just end up resuming it
again, because of the aggressive repetition settings we have for playing
the game.
With this change we introduce the infrastructure for disabling
auto-key-repeat entirely for certain keys. We make use of it for Enter,
Space, and F3, which control game restarts, stone drops, and
the pause/resume state of the game.