Print score summary at end of game
Currently we only print a score summary when we level up. However, it
would be nice to get an overview once the game is over as well.
This change moves the score printing from the Score type itself into the
game and adds additional logic to print the score once the game is over
to address this deficiency.