Introduce State::Quit variant
Our handling of the desire to quit the application is a bit icky: we
assume it can only happen from the key event handler and we furthermore
infer such a desire by taking a peek at the control flow state, which is
rather easy to forget.
With this change we improve on this situation by introducing the
State::Quit variant. If set, the application will exit subsequently.
Having this variant be a first class citizen makes correct handling much
more likely and it means that each part of the game that returns a State
object now has the ability to quit the game.