Introduce Tick enumeration
Our "tick" logic works by returning an Instant to wake up the next time.
Because if nothing is happening there is no need to wake up at all, the
corresponding method on the Game type actually returns an
Option<Instant>.
This is a bit ad-hoc and with upcoming changes we would potentially like
to convey more constraints to callers. As such, this change introduces a
proper enum, Tick, that describes when the next "tick", if any, should
occur. Doing so removes our ghetto min_instant() and maybe_min_instant()
functionality, in favor of just creating an ordering for said enum.