Rename State enum to Change
The name "state" is very common and we intend to use it more throughout
the program (for lack of alternatives). For the existing State enum,
however, we have a similarly apt name with Change.
With this patch we rename it accordingly, freeing up the term "state"
for other anticipated uses down the line.