Introduce Field State enumeration
So far we have treated the Field as having effectively only a single
state that it can be in. However, that's arguably not the full story, as
indicated by, for example, the assertion that the current stone does not
collide with any merged pieces when moving a stone down. This assertion
hinges on an implicit contract with the caller suggesting not to call
this method when a collision has occurred earlier.
This approach is not entirely clean and so with this change we introduce
a dedicated state enumeration to the Field type. Each method using the
stone is now required to inspect the state first (by virtue of the state
containing the stone only in one variant). In the future we anticipate
introducing additional states.