Remove Game::over attribute
This change harnesses the Field type's state member to remove the
Game::over attribute. Basically, the game is always over when the
field's state indicates that there is a collision. Because the Field
acts as the source of truth, the mirroring attribute in the Game can be
seen as superfluous.
Note that it's not 100% super-duper clean to expose the field's state as
we do. However, we will need to expose it anyway for other reasons with
upcoming changes and so we might as well make the best use of it.