Introduce MoveResult::None variant
With upcoming changes we will introduce a phase to the Field during
which completed lines are cleared. While this phase is in progress, any
operations attempting to move the currently active stone will simply not
do anything.
To represent such inaction properly, this change introduces the None
variant to the MoveResult enumeration.