Make Field::move_stone_down_impl() return Change object
Currently the Field::move_stone_down_impl() method always performs some
action that triggers a re-render. With upcoming changes introducing a
time frame at which completed lines are being cleared, that will no
longer be the case.
With this change we adjust the Field::move_stone_down_impl() method to
return a Change object (in addition to the MoveResult) that indicates
whether contents need to be re-rendered.