Only clear completed lines after a delay
Original Tetris variant had an animation that faded out complete lines
after a while, effectively pausing the game for a brief moment whenever
a line was completed. This pause added a nice strategic component to the
game, because there is now a trade-off between clearing as many lines as
possible versus doing so one at a time. The latter can be beneficial
once the field is significantly occupied to increase the time available
to decide best placement.
With this change we introduce such a short pause when clearing lines to
this game as well. To make this happen we add a new state to the Field
type: one when it is clearing lines.