Allow for stone movement during line clearing
Perhaps somewhat counter-intuitively (or perhaps not...), it feels
somewhat awkward to not have the "next" stone move while we briefly
pause to clear completed lines. During high-speed game play, the lack of
reaction to key presses can be extremely confusing.
To fix this short-coming, with this change we actually handle stone
rotations and horizontal movements while clearing complete lines. Note
that we do not, however, allow for down movements -- they make little
sense and would also be somewhat more complex to handle, because a down
movement could lead to additional cleared lines.