Do not create documentation for library
While the program comes with a lib crate that is included in the
published package and theoretically accessible by anyone, it isn't
actually meant to used this way.
We do not want to provide any backwards compatibility guarantees with
respect to the lib part's API surface and to make that somewhat more
clear, with this change we prevent the creation of any and all
rustdoc-style documentation for it.