Only check relevant lines when clearing completed lines
When clearing completed lines from the field, we check all lines present
in the field for whether they are complete. That is arguably a lot of
unnecessary work, because we already know that only lines affected by
the recent stone merge could have become complete.
With this change we carry over the y-range that the stone covered to
minimize the lines we check for being complete.