Correctly update Field state on final collision
When the current stone collides with the field on down movement it is
moved back up and then merged in. Then we take the next stone and reset
it to its initial position. When this reset fails because of another
collision the game is over. However, we missed to update the Field's
state to properly indicate that a collision has occurred, which can lead
to invariant violations down the line.
Fix this problem by updating the Field's state member correctly in this
case.