Improve "change propagation" logic when dropping stones
The Change state propagation logic in the Field::move_stone_down_impl()
method seems fishy, as we only return the most recently reported
Change value. While likely correct, because we would not change the
Field's state to State::Clearing or State::Colliding (the only cases
where an Change::Unchanged would be reported), it's better to not depend
too much on the current implementation at call sites.
To that end, this patch makes sure that we correctly report the
"highest" change seen as part of the various iterations performed.