Remember and render colliding stone
With the recent change to update the Field internal state to "colliding"
when the next stone collides with already dropped pieces when being
moved to its initial position, we effectively stopped rendering this
stone as well when the game is over.
With this change we keep the colliding stone around and be sure to
render it, as we would if it was still user controllable.