Highlight colliding stone
Currently there is not much of an indication that a game is over versus,
say, it being paused. We don't have no special lettering or anything
like that and it that state has to be inferred by noticing that stones
are stacked up to the top.
With this change we make sure to highlight the colliding stone similar
to what we do when we clear a complete line. That does not make it
immediately obvious either, but it definitely is easier noticeable.