Introduce Stonelike trait
With upcoming changes we would like to abstract over what a stone is.
That is, we would like to have a set of methods that every stone-y
entity implements and based on which we can formulate algorithms.
With this patch we introduce the means for doing so: the Stonelike
trait. The trait effectively captures the essence of a stone being
something comprised of "pieces" and provides methods for moving and
rotating them in unison, among other things.