Introduce Fieldlike trait
Similar to what we did for stone objects, this change introduces the
Fieldlike trait as an abstraction over anything that behaves like a
field. In a nutshell, a field is a 2D matrix of pieces that can be
indexed into. It is also able to interact with anything Stonelike in
certain ways, including assimilating the individual pieces and then
identifying "complete" lines and removing them.