Introduce ai module
The one remaining piece that the original C++ version had and our
"clone" does not yet is a self-playing AI.
This change marks the corner stone for the addition of such a AI. It
introduces the ai module and a first basic building block: the
Orientation enum and Orientations struct. These types will allow us to
keep track of the rotation state of a stone relative to its original
orientation. This will become useful once we start searching for valid
moves of a stone and rotations specifically.