Introduce ai::Stone type
Our AI will ultimately be performing a search of what is considered the
best next move of the currently active stone. However, we will not be
using the "regular" representation of a stone as employed by the game.
Rather, the AI will work with its own custom type that does not come
with game specifics such as a texture and may be optimized for the
search down the line.
This change introduces such a stone type to the ai module.