Add tests for Stone and Field
This change adds a bunch of tests for the recently introduced ai::Stone
and ai::Field types. Covered is basic functionality such as stone
rotation or merging of a stone's pieces with a field, along with
clearing complete lines. We use a bunch of utility macros to more
declaratively specify stones and fields for testing, which is expected
to come in even more handy down the line.