Introduce Cost type
Our search for the best action to perform will ultimately be guided by
costs: an estimation of how "expensive" a state (e.g., field) and the
operations to reach to it, are.
We don't want to force us into any particular value range and as such
representing costs as floating points is the most natural choice.
However, floating point values by default don't form a total order,
making them insufficient for our use case.
To work around this problem Rust provided the f32::total_cmp()
functionality. With this change we introduce wrapper around a f32 that
uses this functionality to establish a total order, while offering
sufficiently many operations to make it useful for our intents and
purposes down the line.