Introduce type for remembering "visited" stones
Our search for the best possible action to perform on a stone will
involve moving and rotating stones and checking their "utility" (or
cost). It is basically impossible to know whether a "state" (a stone at
a certain location with a certain rotation) has already been evaluated
in advance and so we need a way to remember the already "visited"
states.
This change introduces the VisitedStones type that can be used to help
with this task. It can be used to remember that a stone has been visited
and the check for the same can be performed quickly.