Introduce State type
Our search for the best set of actions to perform will require some kind
of core data structure that captures the field, the stone being moved,
as well as all the other meta data that we need to make the best
decision in the end.
This change introduces such a data structure with the State type.
Objects of this type form a chain with each instance roughly
representing one action that we intend to perform.