Implement best-action sequence search
This change implements the search algorithm for finding the estimated
best sequence of actions to take. We start with a State that contains
the current Field as well as Stone (and some more meta data). We
"expand" said State by applying all possible actions to it, over and
over until we find a state that has a conflict. We backtrack one step
and should reach the last valid state.
We then use a heuristic function that estimates the "cost" of this state
(really only the field) taking into account the number of "holes", the
"roughness" of the field, the highest piece's position, and more.
The lowest cost state is eventually chosen and a list of actions to get
there constructed.