Hook up auto-playing AI
This change hooks up the auto-playing AI to the main game. Once enabled
(currently controlled by the F2 key), the AI will search for the best
sequence of actions and then perform these as the game advances. The AI
does not "drop" stones, but rather just relies on auto-advancement at
game speed.