Add "stress tests" for auto-playing AI
This change adds a few stress tests for the auto-playing AI that use
field constellations that we struggled in the past to recover from.
These tests are ignored by default because at least at this time they
run forever (and we should probably strive to keep it this way). These
tests are mostly useful as regression tests to see how changes to the
core AI algorithm manifest.

$ cargo test --release -- play_to_fix_mess2 --ignored --nocapture