Build nightly configuration in CI
We use "unit benchmarks" that are currently unstable for our
benchmarking matters. As such, these require a nightly toolchain to
build. That's one of the reasons why so far we have not build said
benchmarks in CI.
With this patch we change that state of affairs, adding a CI build using
the nightly feature using a nightly toolchain. The recent upgrade to
using winit 0.29.2 has broken these builds and so with this change we
also fix this breakage.