Provide proper defaults for game configuration attributes
The way our Config types are defined, serde is only able to deserialize
a complete object with all fields set. That is obviously not great once
we start adding new members, because a mere addition then causes a
deserialization failure of a previously-valid configuration.
With this change we annotate individual attributes with default values
that serde can use, resulting in a much more flexible arrangement: any
value left unspecified by the user in the configuration file will just
be populated with a default. We do that for the overall config and the
game specific part. For the keyboard configuration we preserve
all-or-nothing semantics, to not clutter the code overly much (due to
API constraints, we always query both attributes in tandem).