Remove confusing doc = false setting for tetromino_impl
The doc field of the [lib] section is documented as [0]:
> The doc field indicates whether or not the target is included in the
> documentation generated by cargo doc by default.

While it seems to be doing that locally, that does not seem to be the
case on docs.rs, where the library is what is being documented.
Given that and the fact that moving forward we do intend to expose more
types publicly, this change removes this setting altogether.

[0] https://doc.rust-lang.org/cargo/reference/cargo-targets.html#the-doc-field