Add crate providing xlock bindings
With upcoming changes we would like to provide the means for running
Tetromino as a screen lock. However, we have no intention of
implementing the locking logic ourselves. Rather, we aim to piggy-back
on the existing xlock software distribution and will provide a "mode"
for it. Given that xlock is C software and "modes" are generally closely
tied to the xlock source code, we will need bindings that expose the
necessary types.
This change adds the infrastructure for creating such bindings. It
includes everything from downloading and unpacking the corresponding
source code to invoking bindgen. We check in the bindings for version
5.73 of the program, as a convenience, but it is entirely possible that
those will not work on other versions or when xlock is configured in a
non-standard way.