Expose more types via public API surface
With upcoming changes we intend to introduce an xlock(1) "mode" that
runs Tetromino when a screen lock is active. Doing so requires
instantiation of the game context inside an existing window and in a way
that is more callback based. Luckily our API already provides all the
necessary bits, but not everything needed is publicly accessible yet.
With this patch we adjust the visibility of all types and members that
we need to expose in order to allow for creation of such logic.